+++ src/main/java/org/example/linkshorter/entity/User.java
@@ -9,7 +9,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
     @Column(name = "username")
     private String username;
     @Column(name = "password")
@@ -20,7 +20,7 @@ public class User {
     @JoinColumn(name = "role_id", referencedColumnName = "id")
     private Role role;
 
-    public User(Integer id, String username, String password, String email, Role role) {
+    public User(Long id, String username, String password, String email, Role role) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -31,11 +31,11 @@ public class User {
     public User() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
